product-4: renamed the mapper DTO for the response
