+++ src/main/java/com/zufar/onlinestore/cart/converter/ShoppingSessionItemDtoConverter.java
@@ -2,7 +2,7 @@ package com.zufar.onlinestore.cart.converter;
 
 import com.zufar.onlinestore.cart.dto.ShoppingSessionItemDto;
 import com.zufar.onlinestore.cart.entity.ShoppingSessionItem;
-import com.zufar.onlinestore.product.converter.ProductInfoDtoConverter;
+import com.zufar.onlinestore.product.mapper.ProductInfoDtoConverter;
 import com.zufar.onlinestore.product.dto.ProductInfoFullDto;
 import com.zufar.onlinestore.product.entity.ProductInfo;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +15,7 @@ public class ShoppingSessionItemDtoConverter {
     private final ProductInfoDtoConverter productInfoDtoConverter;
 
     public ShoppingSessionItemDto toDto(final ShoppingSessionItem entity) {
-        ProductInfoFullDto productInfo = productInfoDtoConverter.convertToFullDto(entity.getProductInfo());
+        ProductInfoFullDto productInfo = productInfoDtoConverter.toFullDto(entity.getProductInfo());
 
         return new ShoppingSessionItemDto(
                 entity.getId(),
@@ -26,7 +26,7 @@ public class ShoppingSessionItemDtoConverter {
     }
 
     public ShoppingSessionItem toEntity(final ShoppingSessionItemDto dto) {
-        ProductInfo productInfo = productInfoDtoConverter.convertToEntity(dto.productInfo());
+        ProductInfo productInfo = productInfoDtoConverter.toEntity(dto.productInfo());
 
         return new ShoppingSessionItem(
                 dto.id(),
